+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/screen/home.dart';
+import 'package:flutter_application_1/screen/welcome_screen/welcome.dart';
 
 void main() {
   runApp(MyApp());
@@ -8,13 +8,12 @@ void main() {
 class MyApp extends StatelessWidget {
   @override
   MyApp({Key? key}) : super(key: key);
-  Widget build(BuildContext context){
+  Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home:HomeScreen()
-    );
+        title: 'Flutter Demo',
+        theme: ThemeData(
+          primarySwatch: Colors.green,
+        ),
+        home: WelcomeScreen());
   }
 }
