+++ lib/screen/constant/style.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+//Color
+const Color kcPrimaryColor = Color.fromARGB(255, 161, 227, 178);
+const Color kcMediumGreyColor = Color(0xff868686);
+
+//Font Sizing
+const double kBodyTextSize = 16;
+
+//TextStyle
+const TextStyle ktsMediumGreyBodyText = TextStyle(
+  color: kcMediumGreyColor,
+  fontSize: kBodyTextSize,
+);
