+++ lib/screen/constant/ui_helper.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+//Horizontal Spacing
+const Widget horizontalSpaceTiny = SizedBox(width: 5.0);
+const Widget horizontalSpaceSmall = SizedBox(width: 10.0);
+const Widget horizontalSpaceRegular = SizedBox(width: 18.0);
+const Widget horizontalSpaceMedium = SizedBox(width: 25.0);
+const Widget horizontalSpaceLarge = SizedBox(width: 50.0);
+
+//Vertical Spacing
+const Widget verticalSpaceTiny = SizedBox(width: 5.0);
+const Widget verticalSpaceSmall = SizedBox(width: 10.0);
+const Widget verticalSpaceRegular = SizedBox(width: 18.0);
+const Widget verticalSpaceMedium = SizedBox(width: 25.0);
+const Widget verticalSpaceLarge = SizedBox(width: 50.0);
