+++ lib/screen/register.dart
@@ -4,8 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_application_1/model/profile.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:form_field_validator/form_field_validator.dart';
-
-import 'home.dart';
+import 'welcome_screen/welcome.dart';
 
 class RegisterScreen extends StatefulWidget {
   const RegisterScreen({Key? key}) : super(key: key);
@@ -91,7 +90,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
                                           gravity: ToastGravity.TOP);
                                       Navigator.pushReplacement(context,
                                           MaterialPageRoute(builder: (context) {
-                                        return HomeScreen();
+                                        return WelcomeScreen();
                                       }));
                                     });
                                   } on FirebaseAuthException catch (e) {
